% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corr_comb}
\alias{corr_comb}
\title{Corrected starch counts combined}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 47 rows and 9 columns.
}
\usage{
corr_comb
}
\description{
Corrected starch counts combined
}
\keyword{datasets}
