% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corr_counts}
\alias{corr_counts}
\title{Corrected starch counts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58 rows and 10 columns.
}
\usage{
corr_counts
}
\description{
Corrected starch counts
}
\keyword{datasets}
