% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{correct_count}
\alias{correct_count}
\title{Correct starch counts}
\usage{
correct_count(raw, prop, vol)
}
\arguments{
\item{raw}{integer. absolute number of starches counted.}

\item{prop}{numeric. Proportion of slide that was counted.}

\item{vol}{numeric. Volume of EDTA on the mounted slide.}
}
\description{
Calculates the number of starches on a whole microscope slide based on the
proportion of slide that was actually counted.
}
