% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{extrap_count}
\alias{extrap_count}
\title{Extrapolates the counts from a slide to the total count over the course of the
experiment.}
\usage{
extrap_count(raw, prop, vol, days)
}
\arguments{
\item{vol}{numeric. Volume of starch added to the sample wells.}

\item{days}{integer. Number of days starch was included in experiment treatments.}

\item{@inheritParams}{correct_count}
}
\description{
Extrapolates the counts from a slide to the total count over the course of the
experiment.
}
