% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amylase.R
\name{sample_calc}
\alias{sample_calc}
\title{Amylase activity in sample calculation}
\usage{
sample_calc(samp_data, cols = NULL, std_curve, dilf, ml_enz)
}
\arguments{
\item{samp_data}{data frame with the photometric readings of the samples.}

\item{cols}{vector containing the column numbers of the sample photometric readings. If NULL, all columns assumed to be standard curve measurements.}

\item{std_curve}{an object from the \code{stdcurve} function output.}

\item{dilf}{dilution factor of samples. Or, dad I'd like to ...}

\item{ml_enzyme}{volume of the sample used in the assay.}
}
\description{
Amylase activity in sample calculation
}
