% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amylase.R
\name{sep_samples}
\alias{sep_samples}
\title{Highly specialised function to separate saliva and bmm samples}
\usage{
sep_samples(data)
}
\arguments{
\item{data}{data frame containing the photometric readings}
}
\description{
Highly specialised function to separate saliva and bmm samples
}
